Add Keyword Tests to WhenBuildingKeywordLink
