+++ src/SFA.DAS.FAT.Web.UnitTests/Models/CoursesViewModelTests/WhenBuildingKeywordLink.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using AutoFixture.NUnit3;
+using NUnit.Framework;
+using System;
 using System.Collections.Generic;
 using System.Text;
 
@@ -6,5 +8,50 @@ namespace SFA.DAS.FAT.Web.UnitTests.Models.CoursesViewModel
 {
     class WhenBuildingKeywordLink
     {
+        [Test, AutoData]
+        public void Then_The_Clear_Keyword_Link_Is_Generated_If_Filtered_By_Keyword(string keyword)
+        {
+            //Arrange Act
+            var model = new Web.Models.CoursesViewModel
+            {
+                Keyword = keyword,
+            };
+
+            //Assert
+            Assert.IsNotNull(model.ClearKeywordLink);
+            Assert.AreEqual("", model.ClearKeywordLink);
+        }
+
+        [Test, AutoData]
+        public void Then_The_Clear_Keyword_Link_Is_Generated_If_Filtered_By_Keyword_With_Sectors(List<Guid> selectedRoutes, string keyword)
+        {
+            //Arrange Act
+            var model = CoursesViewModelTests.CoursesViewModelFactory.BuildModel(selectedRoutes, keyword, new List<int>());
+
+            //Assert
+            Assert.IsNotNull(model.ClearKeywordLink);
+            Assert.AreEqual("?sectors=" + string.Join("&sectors=", model.SelectedSectors), model.ClearKeywordLink);
+        }
+
+        [Test, AutoData]
+        public void Then_The_Clear_Keyword_Link_Is_Generated_If_Filtered_By_Keyword_With_Levels(List<int> selectedLevels, string keyword)
+        {
+            //Arrange Act
+            var model = CoursesViewModelTests.CoursesViewModelFactory.BuildModel(new List<Guid>(), keyword, selectedLevels);
+
+            //Assert
+            Assert.IsNotNull(model.ClearKeywordLink);
+            Assert.AreEqual("?levels=" + string.Join("&levels=", model.SelectedLevels), model.ClearKeywordLink);
+        }
+        [Test, AutoData]
+        public void Then_The_Clear_Keyword_Link_Is_Generated_If_Filtered_By_Keyword_With_Sectors_And_Levels(List<Guid> selectedRoutes, List<int> selectedLevels, string keyword)
+        {
+            //Arrange Act
+            var model = CoursesViewModelTests.CoursesViewModelFactory.BuildModel(selectedRoutes, keyword, selectedLevels);
+
+            //Assert
+            Assert.IsNotNull(model.ClearKeywordLink);
+            Assert.AreEqual("?sectors=" + string.Join("&sectors=", model.SelectedSectors) + "&levels=" + string.Join("&levels=", model.SelectedLevels), model.ClearKeywordLink);
+        }
     }
 }
